Move message.ts
