+++ apps/website/src/app/main/app/app.component.ts
@@ -2,7 +2,7 @@ import {Component, OnInit} from '@angular/core'
 import {SwUpdate} from '@angular/service-worker'
 import {MessageService} from '../../services/error/message.service'
 import {ThemeService} from '../../services/theme/theme.service';
-import {Message} from '../../messages/Message';
+import {Message} from '../../shared/types/Message';
 import {swipeRight} from '@angulon/ui';
 import * as AOS from 'aos';
